move model files to CE
