+++ apps/meteor/server/models/WorkspaceCredentials.ts
@@ -1,6 +1,6 @@
 import { registerModel } from '@rocket.chat/models';
 
-import { db } from '../../../server/database/utils';
+import { db } from '../database/utils';
 import { WorkspaceCredentialsRaw } from './raw/WorkspaceCredentials';
 
 registerModel('IWorkspaceCredentialsModel', new WorkspaceCredentialsRaw(db));
